Separated mmu_sync_feedback class; bug fixes
